+++ gimmemotifs/orthologs.py
@@ -44,7 +44,9 @@ def motif2factor_from_orthologs(
     database_references: List[str] = None,
     extra_orthologs_references: List[str] = None,
     new_reference: List[str] = None,
+    genomes_dir: str = None,
     tmpdir: str = None,
+    keep_intermediate: bool = False,
     outdir: str = ".",
     strategy: str = "lenient",
     threads: int = 24,
@@ -79,7 +81,10 @@ def motif2factor_from_orthologs(
 
     _check_install()
 
+    # Check if we can write output before we do a lot of work
+    os.makedirs(outdir, exist_ok=True)
     tmpdir = tempfile.mkdtemp() if tmpdir is None else tmpdir
+    genomes_dir = get_genomes_dir(genomes_dir, check_exist=False)
 
     logger.info(f"Making a new reference for: {' & '.join(new_reference)}.")
     logger.info(
@@ -88,19 +93,22 @@ def motif2factor_from_orthologs(
     logger.info(
         f"For better orthology inference we are also using these assemblies: {' & '.join(extra_orthologs_references)}."
     )
-    logger.info(f"Using strategy: {strategy} for orthology/name inference.")
-    logger.info(f"tmpdir: {tmpdir}.")
-    logger.info(f"outdir: {outdir}.")
-
-    all_genomes = set(database_references + extra_orthologs_references + new_reference)
+    logger.info(f"Using {strategy} strategy for orthology/name inference.")
+    logger.info(f"genomes_dir: {genomes_dir}")
+    logger.info(f"tmpdir: {tmpdir}")
+    logger.info(f"outdir: {outdir}")
 
     # download all required genomes
     logger.info("Downloading all assemblies.")
-    _download_genomes_with_annot(all_genomes, tmpdir)
+    all_genomes = list(
+        set(database_references + extra_orthologs_references + new_reference)
+    )
+    all_genomes = _download_genomes_with_annot(all_genomes, genomes_dir, tmpdir)
 
     # convert each genome + annotation into the primary genes (longest protein per gene)
     logger.info("Taking the longest protein per gene per assembly.")
     for genome in all_genomes:
+        logger.info(f"Processing {genome}.")
         annot2primpep(genome, tmpdir)
 
     # run orthofinder on our primary genes
@@ -121,7 +129,10 @@ def motif2factor_from_orthologs(
 
     # process the references
     logger.info("Now writing your motif2factors files.")
-    for genome in new_reference:
+    # make sure the output dir exists
+    pathlib.Path(f"{outdir}").mkdir(parents=True, exist_ok=True)
+    for reference in new_reference:
+        genome = genomepy.Genome(reference, genomes_dir, build_index=False).name
         make_motif2factors(
             f"{outdir}/{genome}.{database}",
             new_reference=genome,
@@ -133,7 +144,8 @@ def motif2factor_from_orthologs(
         )
 
     # cleanup
-    shutil.rmtree(tmpdir)
+    if not keep_intermediate:
+        shutil.rmtree(tmpdir)
 
 
 def _check_install():
@@ -154,7 +166,9 @@ def _orthofinder(peptide_folder, threads):
     """
     # run orthofinder on the primary transcripts
     result = subprocess.run(
-        ["orthofinder", "-f", peptide_folder, "-t", str(threads)], capture_output=True
+        ["orthofinder", "-f", peptide_folder, "-t", str(threads)],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
     )
 
     logger.debug(f"""stdout of orthofinder:\n {result.stdout.decode("utf-8")}""")
@@ -166,46 +180,103 @@ def _orthofinder(peptide_folder, threads):
     return orthofinder_result
 
 
-def _download_genomes_with_annot(genomes, genomes_dir):
-    # make sure each genome has a gene annotation
-    no_annotations = [genome for genome in genomes if not _has_annotation(genome)]
-    assert (
-        len(no_annotations) == 0
-    ), f"genome(s): {','.join(no_annotations)} seem not to have an annotation for it."
+def _prepare_genomes_with_annot(genome, genome_dir, outdir):
+    # unzip genome and gtf (if needed) and symlink these in the tmpdir
+    fasta = os.path.join(genome_dir, f"{genome}.fa")
+    anno = os.path.join(genome_dir, f"{genome}.annotation.gtf")
+    for gp_file in [fasta, anno]:
+        # unzip files if needed
+        if not os.path.exists(gp_file):
+            gzipped_gp_file = gp_file + ".gz"
+            result = subprocess.run(
+                ["gunzip", gzipped_gp_file],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+            )
+            logger.debug(f"""stdout of gunzip:\n {result.stdout.decode("utf-8")}""")
+            logger.debug(f"""stderr of gunzip:\n {result.stderr.decode("utf-8")}""")
+        # symlink files in tmpdir
+        tmp_file = os.path.join(outdir, f"{os.path.basename(gp_file)}")
+        if os.path.exists(tmp_file):
+            # in case of a (partial) rerun
+            os.unlink(tmp_file)
+        os.symlink(
+            gp_file,
+            tmp_file,
+        )
 
-    # download the genomes
-    # add check to see if not already in genomes dir?
-    default_genomes_dir = get_genomes_dir(check_exist=False)
-    logger.debug(f"using default genome dir: {default_genomes_dir}")
-    for genome in genomes:
+
+def _download_genomes_with_annot(genomes, genomes_dir, tpmdir):
+    """
+    Download missing genomes/annotations.
+    Accepts genome names, genomepy genome directories, and fasta filepaths.
+    Will use existing genomes where available.
+
+    Returns
+    -------
+    genomes : list
+        a list of genome names, filepaths & suffixes removed (if any)
+    """
+    existing_genomes = []
+    for i, genome in enumerate(genomes):
+        try:
+            gp = genomepy.Genome(genome, genomes_dir)
+            genomes[i] = genome = gp.name  # filepath -> name
+            genome_dir = gp.genome_dir
+        except FileNotFoundError:
+            genome_dir = os.path.join(genomes_dir, genome)
+        outdir = os.path.join(tpmdir, genome)
+        os.makedirs(outdir, exist_ok=True)
+
+        # first check if the user supplied a pep.fa
+        pep = os.path.join(genome_dir, f"{genome}.pep.fa")
+        tmp_pep = os.path.join(outdir, f"{genome}.pep.fa")
+        if os.path.exists(pep):
+            logger.info(f"found pep.fa for {genome}, using that one.")
+            if os.path.exists(tmp_pep):
+                # in case of a (partial) rerun
+                os.unlink(tmp_pep)
+            os.symlink(
+                pep,
+                tmp_pep,
+            )
+            existing_genomes.append(genome)
+            continue
+
         # check if already in default genomes dir, if so, skip downloading and directly copy
+        required = [
+            os.path.join(genome_dir, f"{genome}.{ext}")
+            for ext in ["fa", "annotation.gtf"]
+        ]
         if all(
-            os.path.exists(f"{default_genomes_dir}/{genome}/{genome}.{extension}")
-            for extension in ["fa", "annotation.gtf"]
+            os.path.exists(file) or os.path.exists(f"{file}.gz") for file in required
         ):
-            logger.info(f"{genome} was already downloaded, using that version.")
-            # if except, probably a continuation from previous run
+            # if except, probably a continuation from previous run or NTFS related issues
             try:
-                os.mkdir(f"{genomes_dir}/{genome}")
-                os.symlink(
-                    f"{default_genomes_dir}/{genome}/{genome}.fa",
-                    f"{genomes_dir}/{genome}/{genome}.fa",
-                )
-                os.symlink(
-                    f"{default_genomes_dir}/{genome}/{genome}.annotation.gtf",
-                    f"{genomes_dir}/{genome}/{genome}.annotation.gtf",
-                )
+                _prepare_genomes_with_annot(genome, genome_dir, outdir)
+                existing_genomes.append(genome)
+                logger.info(f"{genome} was already downloaded, using that version.")
             except Exception:
                 pass
-        else:
-            logger.info(f"Downloading {genome} through genomepy.")
-            genomepy.install_genome(genome, annotation=True, genomes_dir=genomes_dir)
-            result = subprocess.run(
-                ["gunzip", f"{genomes_dir}/{genome}/{genome}.annotation.gtf.gz"],
-                capture_output=True,
-            )
-            logger.debug(f"""stdout of genomepy:\n {result.stdout.decode("utf-8")}""")
-            logger.debug(f"""stderr of genomepy:\n {result.stderr.decode("utf-8")}""")
+
+    # make sure each genome to download has a gene annotation
+    download_genomes = [genome for genome in genomes if genome not in existing_genomes]
+    no_annotations = [
+        genome for genome in download_genomes if not _has_annotation(genome)
+    ]
+    assert (
+        len(no_annotations) == 0
+    ), f"genome(s): {','.join(no_annotations)} do not seem to have a gene annotation available!"
+
+    # download missing genomes, or genomes with missing annotation
+    for genome in download_genomes:
+        logger.info(f"Downloading {genome} through genomepy.")
+        genomepy.install_genome(genome, annotation=True, genomes_dir=genomes_dir)
+        genome_dir = os.path.join(genomes_dir, genome)
+        outdir = os.path.join(tpmdir, genome)
+        _prepare_genomes_with_annot(genome, genome_dir, outdir)
+
+    return genomes  # all names clean
 
 
 def annot2primpep(genome, outdir):
@@ -223,24 +294,26 @@ def annot2primpep(genome, outdir):
     # setup our result folder
     pathlib.Path(f"{outdir}/prim_genes").mkdir(parents=True, exist_ok=True)
 
-    # for each genome, make a .pep.fa. This pep.fa contains the LONGEST protein for each gene
-    # use gffread to convert our annotation.gtf into all possible peptides
-    result = subprocess.run(
-        [
-            "gffread",
-            "-y",
-            f"{outdir}/{genome}/{genome}.pep.fa",
-            "-g",
-            f"{outdir}/{genome}/{genome}.fa",
-            f"{outdir}/{genome}/{genome}.annotation.gtf",
-            "-S",
-            "--table",
-            "gene_name,gene_id",
-        ],
-        capture_output=True,
-    )
-    logger.debug(f"""stdout of gffread:\n {result.stdout.decode("utf-8")}""")
-    logger.debug(f"""stderr of gffread:\n {result.stderr.decode("utf-8")}""")
+    if not os.path.exists(f"{outdir}/{genome}/{genome}.pep.fa"):
+        # for each genome, make a .pep.fa. This pep.fa contains the LONGEST protein for each gene
+        # use gffread to convert our annotation.gtf into all possible peptides
+        result = subprocess.run(
+            [
+                "gffread",
+                "-y",
+                f"{outdir}/{genome}/{genome}.pep.fa",
+                "-g",
+                f"{outdir}/{genome}/{genome}.fa",
+                f"{outdir}/{genome}/{genome}.annotation.gtf",
+                "-S",
+                "--table",
+                "gene_name,gene_id",
+            ],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+        )
+        logger.debug(f"""stdout of gffread:\n {result.stdout.decode("utf-8")}""")
+        logger.debug(f"""stderr of gffread:\n {result.stderr.decode("utf-8")}""")
 
     # read the resulting .pep.fa
     proteins = pyfaidx.Fasta(f"{outdir}/{genome}/{genome}.pep.fa", read_long_names=True)
@@ -248,11 +321,15 @@ def annot2primpep(genome, outdir):
     # and only keep the longest protein per gene
     records = dict()
     for record in proteins:
-        # get the gene name and gene id of the protein
-        prot_name = "|".join(record.long_name.split("\t")[1:])
+        # get the gene name and gene id of the protein (last two identifiers)
+        prot_name = re.split(r"[\s|~]+", record.long_name)[-2:]
+        # if we have only one identifier (user-defined pep.fa) just add a dot to indicate missing
+        if len(prot_name) == 1:
+            prot_name.append(".")
+        prot_name = "|".join(prot_name)
 
-        # get the protein sequence
-        protein = str(record)
+        # get the protein sequence (remove ending stop codon if present)
+        protein = str(record).strip("*")
 
         # skip proteins with stop codon
         # (probably mitochondrial protein since they have a different codon table)
@@ -449,7 +526,7 @@ def make_motif2factors(
 
     with open(f"{prefix}.pfm", "w") as f:
         for motif in motifs:
-            print(motif.to_pwm(), file=f)
+            print(motif.to_ppm(), file=f)
 
 
 @lru_cache(maxsize=99999)
@@ -580,12 +657,14 @@ def _has_annotation(genome):
     """
     Return True if a genome has an annotation for it, else False
     """
-    providers = [
-        genomepy.ProviderBase.create(provider)
-        for provider in ["ensembl", "ucsc", "ncbi"]
-    ]
+    if hasattr(genomepy, "Provider"):
+        base = genomepy.Provider
+    else:
+        base = genomepy.ProviderBase
+    providers = [base.create(provider) for provider in ["ensembl", "ucsc", "ncbi"]]
     for provider in providers:
         if genome in provider.genomes:
             link = provider.get_annotation_download_link(genome)
             return link is not None
+    logger.warning(f"No annotation found for {genome}")
     return False