+++ gimmemotifs/tools/gadem.py
@@ -117,7 +117,7 @@ class Gadem(MotifProgram):
 
             motifs.append(Motif(pwm))
             motifs[-1].id = "{}_{}".format(self.name, m_id)
-            # motifs[-1].pwm = pwm
+
             if align:
                 motifs[-1].pfm = pfm
                 motifs[-1].align = align